Hooks folder and commented/condensed
